refactor: move plan.rs into plan/ module with separate tests (#1376)
Split the monolithic plan module into src/plan/mod.rs (types + Operation)
and src/plan/tests.rs. Update reference-doc inventory test to read
plan/mod.rs.

Refs #1376
Refs #1372
